pyproject.toml
[project]
name = "pydreamtv"
version = "0.1.0"
description = "Stream IPTV to Discord"
authors = [
    {name = "Chris Read",email = "centurix@gmail.com"},
]
dependencies = [
    "pyvirtualcam>=0.10.2",
    "opencv-python>=4.8.0.76",
    "requests>=2.31.0",
    "mpv>=1.0.4",
    "lxml>=4.9.3",
    "mysql-connector-python>=8.1.0",
    "click>=8.1.7",
    "discord-py>=2.3.2",
    "pillow>=10.0.0",
    "python-dotenv>=1.0.0",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
